Struct constellation::constellation::Frame

struct Frame {
    id: FrameId,
    prev: Vec<FrameState>,
    current: FrameState,
    next: Vec<FrameState>,
}

A frame in the frame tree. Each frame is the constrellation's view of a browsing context. Each browsing context has a session history, caused by navigation and traversing the history. Each frame has its current entry, plus past and future entries. The past is sorted chronologically, the future is sorted reverse chronoogically: in partiucular prev.pop() is the latest past entry, and next.pop() is the earliest future entry.

Fields
id: FrameId

The frame id.

prev: Vec<FrameState>

The past session history, ordered chronologically.

current: FrameState

The currently active session history entry.

next: Vec<FrameState>

The future session history, ordered reverse chronologically.

Methods

impl Frame

fn new(id: FrameId, pipeline_id: PipelineId) -> Frame

Create a new frame. Note this just creates the frame, it doesn't add it to the frame tree.

fn load(&mut self, pipeline_id: PipelineId)

Set the current frame entry, and push the current frame entry into the past.

fn remove_forward_entries(&mut self) -> Vec<FrameState>

Set the future to be empty.

fn replace_current(&mut self, pipeline_id: PipelineId) -> FrameState

Set the current frame entry, and drop the current frame entry.