Struct constellation::constellation::InitialConstellationState

pub struct InitialConstellationState {

    pub embedder_proxy: EmbedderProxy,
    pub compositor_proxy: CompositorProxy,
    pub devtools_chan: Option<Sender<DevtoolsControlMsg>>,
    pub bluetooth_thread: IpcSender<BluetoothRequest>,
    pub font_cache_thread: FontCacheThread,
    pub public_resource_threads: ResourceThreads,
    pub private_resource_threads: ResourceThreads,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub webrender_document: DocumentId,
    pub webrender_api_sender: RenderApiSender,
    pub webrender_external_images: Arc<Mutex<WebrenderExternalImageRegistry>>,
    pub webgl_threads: Option<WebGLThreads>,
    pub webxr_registry: Registry,
    pub glplayer_threads: Option<GLPlayerThreads>,
    pub player_context: WindowGLContext,
    pub event_loop_waker: Option<Box<dyn EventLoopWaker>>,
    pub user_agent: Cow<'static, str>,
    pub wgpu_image_map: Arc<Mutex<HashMap<u64, PresentationData>>>,

}

State needed to construct a constellation.

Fields
embedder_proxy: EmbedderProxy

A channel through which messages can be sent to the embedder.

compositor_proxy: CompositorProxy

A channel through which messages can be sent to the compositor.

devtools_chan: Option<Sender<DevtoolsControlMsg>>

A channel to the developer tools, if applicable.

bluetooth_thread: IpcSender<BluetoothRequest>

A channel to the bluetooth thread.

font_cache_thread: FontCacheThread

A channel to the font cache thread.

public_resource_threads: ResourceThreads

A channel to the resource thread.

private_resource_threads: ResourceThreads

A channel to the resource thread.

time_profiler_chan: ProfilerChan

A channel to the time profiler thread.

mem_profiler_chan: ProfilerChan

A channel to the memory profiler thread.

webrender_document: DocumentId

Webrender document ID.

webrender_api_sender: RenderApiSender

Webrender API.

webrender_external_images: Arc<Mutex<WebrenderExternalImageRegistry>>

Webrender external images

webgl_threads: Option<WebGLThreads>

Entry point to create and get channels to a WebGLThread.

webxr_registry: Registry

The XR device registry

glplayer_threads: Option<GLPlayerThreads>player_context: WindowGLContext

Application window’s GL Context for Media player

event_loop_waker: Option<Box<dyn EventLoopWaker>>

Mechanism to force the compositor to process events.

user_agent: Cow<'static, str>

User agent string to report in network requests.

wgpu_image_map: Arc<Mutex<HashMap<u64, PresentationData>>>