constellation::constellation_webview

Struct ConstellationWebView

pub(crate) struct ConstellationWebView {
    webview_id: WebViewId,
    pub focused_browsing_context_id: BrowsingContextId,
    pub hovered_browsing_context_id: Option<BrowsingContextId>,
    pub last_mouse_move_point: Point2D<f32, CSSPixel>,
    pub session_history: JointSessionHistory,
    pub user_content_manager_id: Option<UserContentManagerId>,
    theme: Theme,
}

The Constellation’s view of a WebView in the embedding layer. This tracks all of the Constellation state for this WebView.

Fields
webview_id: WebViewId

The WebViewId of this ConstellationWebView.

focused_browsing_context_id: BrowsingContextId

The currently focused browsing context in this webview for key events. The focused pipeline is the current entry of the focused browsing context.

hovered_browsing_context_id: Option<BrowsingContextId>

The BrowsingContextId of the currently hovered browsing context, to use for knowing which frame is currently receiving cursor events.

last_mouse_move_point: Point2D<f32, CSSPixel>

The last mouse move point in the coordinate space of the Pipeline that it happened int.

session_history: JointSessionHistory

The joint session history for this webview.

user_content_manager_id: Option<UserContentManagerId>

The UserContentManagerId for all pipelines in this WebView. This is Some if the embedder has set a UserContentManager using the WebViewBuilder API and it is None otherwise.

theme: Theme

The Theme that this ConstellationWebView uses. This is communicated to all ScriptThreads so that they know how to render the contents of a particular `WebView.

Implementations

impl ConstellationWebView

pub(crate) fn new( webview_id: WebViewId, focused_browsing_context_id: BrowsingContextId, user_content_manager_id: Option<UserContentManagerId>, ) -> Self

pub(crate) fn set_theme(&mut self, new_theme: Theme) -> bool

Set the Theme on this ConstellationWebView returning true if the theme changed.

pub(crate) fn theme(&self) -> Theme

Get the Theme of this ConstellationWebView.

fn target_pipeline_id_for_input_event( &self, event: &ConstellationInputEvent, browsing_contexts: &FxHashMap<BrowsingContextId, BrowsingContext>, ) -> Option<PipelineId>

pub(crate) fn forward_input_event( &mut self, event: ConstellationInputEvent, pipelines: &FxHashMap<PipelineId, Pipeline>, browsing_contexts: &FxHashMap<BrowsingContextId, BrowsingContext>, ) -> bool

Forward the InputEvent to this ConstellationWebView. Returns false if the event could not be forwarded or true otherwise.