Module constellation::frame

Structs

Frame
A frame in the frame tree. Each frame is the constellation's view of a browsing context. Each browsing context has a session history, caused by navigation and traversing the history. Each frame has its current entry, plus past and future entries. The past is sorted chronologically, the future is sorted reverse chronologically: in particular prev.pop() is the latest past entry, and next.pop() is the earliest future entry.

FrameChange
Represents a pending change in the frame tree, that will be applied once the new pipeline has loaded and completed initial layout / paint.

FrameState
An entry in a frame's session history. Each entry stores the pipeline id for a document in the session history.

FrameTreeIterator
An iterator over a frame tree, returning the fully active frames in depth-first order. Note that this iterator only returns the fully active frames, that is ones where every ancestor frame is in the currently active pipeline of its parent frame.

FullFrameTreeIterator
An iterator over a frame tree, returning all frames in depth-first order. Note that this iterator returns all frames, not just the fully active ones.