Struct constellation::frame::FrameChange

pub struct FrameChange {
    pub frame_id: FrameId,
    pub old_pipeline_id: Option<PipelineId>,
    pub new_pipeline_id: PipelineId,
    pub url: ServoUrl,
    pub replace: Option<FrameState>,
}

Represents a pending change in the frame tree, that will be applied once the new pipeline has loaded and completed initial layout / paint.

Fields
frame_id: FrameId

The frame to change.

old_pipeline_id: Option<PipelineId>

The pipeline that was currently active at the time the change started. TODO: can this field be removed?

new_pipeline_id: PipelineId

The pipeline for the document being loaded.

url: ServoUrl

The URL for the document being loaded.

replace: Option<FrameState>

Is the new document replacing the current document (e.g. a reload) or pushing it into the session history (e.g. a navigation)?