Module constellation::pipeline

Structs

InitialPipelineState

Initial setup data needed to construct a pipeline.

NewPipeline

Pipeline

A Pipeline is the constellation’s view of a Window. Each pipeline has an event loop (executed by a script thread). A script thread may be responsible for many pipelines.

UnprivilegedPipelineContent

Creating a new pipeline may require creating a new event loop. This is the data used to initialize the event loop. TODO: simplify this, and unify it with InitialPipelineState if possible.