Struct constellation::pipeline::Pipeline

pub struct Pipeline {
    pub id: PipelineId,
    pub frame_id: FrameId,
    pub parent_info: Option<(PipelineId, FrameType)>,
    pub script_chan: Rc<ScriptChan>,
    pub layout_chan: IpcSender<LayoutControlMsg>,
    pub compositor_proxy: Box<CompositorProxy + 'static + Send>,
    pub url: ServoUrl,
    pub title: Option<String>,
    pub size: Option<TypedSize2D<f32, PagePx>>,
    pub running_animations: bool,
    pub children: Vec<FrameId>,
    pub is_private: bool,
    pub visible: bool,
}

A uniquely-identifiable pipeline of script thread, layout thread, and paint thread.

Fields
id: PipelineId frame_id: FrameId

The ID of the frame that contains this Pipeline.

parent_info: Option<(PipelineId, FrameType)> script_chan: Rc<ScriptChan> layout_chan: IpcSender<LayoutControlMsg>

A channel to layout, for performing reflows and shutdown.

compositor_proxy: Box<CompositorProxy + 'static + Send>

A channel to the compositor.

url: ServoUrl

URL corresponding to the most recently-loaded page.

title: Option<String>

The title of the most recently-loaded page.

size: Option<TypedSize2D<f32, PagePx>> running_animations: bool

Whether this pipeline is currently running animations. Pipelines that are running animations cause composites to be continually scheduled.

children: Vec<FrameId> is_private: bool

Whether this pipeline is considered distinct from public pipelines.

visible: bool

Whether this pipeline should be treated as visible for the purposes of scheduling and resource management.

Methods

impl Pipeline

fn spawn<Message, LTF, STF>(state: InitialPipelineState)
-> Result<(Pipeline, Option<ChildProcess>), IOError> where LTF: LayoutThreadFactory<Message=Message>,
STF: ScriptThreadFactory<Message=Message>

Starts a paint thread, layout thread, and possibly a script thread, in a new process if requested.

fn new(id: PipelineId,
frame_id: FrameId,
parent_info: Option<(PipelineId, FrameType)>,
script_chan: Rc<ScriptChan>,
layout_chan: IpcSender<LayoutControlMsg>,
compositor_proxy: Box<CompositorProxy + 'static + Send>,
is_private: bool,
url: ServoUrl,
size: Option<TypedSize2D<f32, PagePx>>,
visible: bool)
-> Pipeline

fn exit(&self)

fn freeze(&self)

fn thaw(&self)

fn force_exit(&self)

fn to_sendable(&self) -> CompositionPipeline

fn add_child(&mut self, frame_id: FrameId)

fn remove_child(&mut self, frame_id: FrameId)

fn trigger_mozbrowser_event(&self,
child_id: Option<FrameId>,
event: MozBrowserEvent)

fn notify_visibility(&self)

fn change_visibility(&mut self, visible: bool)