Struct constellation::Constellation

pub struct Constellation<Message, LTF, STF> {
    script_sender: IpcSender<FromScriptMsg>,
    script_receiver: Receiver<Result<FromScriptMsg, IpcError>>,
    layout_sender: IpcSender<FromLayoutMsg>,
    layout_receiver: Receiver<Result<FromLayoutMsg, IpcError>>,
    compositor_receiver: Receiver<FromCompositorMsg>,
    compositor_proxy: Box<CompositorProxy>,
    public_resource_threads: ResourceThreads,
    private_resource_threads: ResourceThreads,
    image_cache_thread: ImageCacheThread,
    font_cache_thread: FontCacheThread,
    debugger_chan: Option<Sender>,
    devtools_chan: Option<Sender<DevtoolsControlMsg>>,
    bluetooth_thread: IpcSender<BluetoothRequest>,
    swmanager_chan: Option<IpcSender<ServiceWorkerMsg>>,
    swmanager_sender: IpcSender<SWManagerMsg>,
    swmanager_receiver: Receiver<Result<SWManagerMsg, IpcError>>,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    scheduler_chan: IpcSender<TimerEventRequest>,
    webrender_api_sender: RenderApiSender,
    event_loops: HashMap<FrameId, HashMap<String, Weak<EventLoop>>>,
    pipelines: HashMap<PipelineId, Pipeline>,
    frames: HashMap<FrameId, Frame>,
    pending_frames: Vec<FrameChange>,
    root_frame_id: FrameId,
    focus_pipeline_id: Option<PipelineId>,
    next_pipeline_namespace_id: PipelineNamespaceId,
    window_size: WindowSizeData,
    webdriver: WebDriverData,
    document_states: HashMap<PipelineId, DocumentState>,
    shutting_down: bool,
    handled_warnings: VecDeque<(Option<String>, String)>,
    random_pipeline_closure: Option<(ServoRng, f32)>,
    phantom: PhantomData<(Message, LTF, STF)>,
    webvr_thread: Option<IpcSender<WebVRMsg>>,
}

The Constellation itself. In the servo browser, there is one constellation, which maintains all of the browser global data. In embedded applications, there may be more than one constellation, which are independent of each other.

The constellation may be in a different process from the pipelines, and communicates using IPC.

It is parameterized over a LayoutThreadFactory and a ScriptThreadFactory (which in practice are implemented by LayoutThread in the layout crate, and ScriptThread in the script crate). Script and layout communicate using a Message type.

Fields
script_sender: IpcSender<FromScriptMsg>

An IPC channel for script threads to send messages to the constellation. This is the script threads' view of script_receiver.

script_receiver: Receiver<Result<FromScriptMsg, IpcError>>

A channel for the constellation to receive messages from script threads. This is the constellation's view of script_sender.

layout_sender: IpcSender<FromLayoutMsg>

An IPC channel for layout threads to send messages to the constellation. This is the layout threads' view of layout_receiver.

layout_receiver: Receiver<Result<FromLayoutMsg, IpcError>>

A channel for the constellation to receive messages from layout threads. This is the constellation's view of layout_sender.

compositor_receiver: Receiver<FromCompositorMsg>

A channel for the constellation to receive messages from the compositor thread.

compositor_proxy: Box<CompositorProxy>

A channel (the implementation of which is port-specific) for the constellation to send messages to the compositor thread.

public_resource_threads: ResourceThreads

Channels for the constellation to send messages to the public resource-related threads. There are two groups of resource threads: one for public browsing, and one for private browsing.

private_resource_threads: ResourceThreads

Channels for the constellation to send messages to the private resource-related threads. There are two groups of resource threads: one for public browsing, and one for private browsing.

image_cache_thread: ImageCacheThread

A channel for the constellation to send messages to the image cache thread.

font_cache_thread: FontCacheThread

A channel for the constellation to send messages to the font cache thread.

debugger_chan: Option<Sender>

A channel for the constellation to send messages to the debugger thread.

devtools_chan: Option<Sender<DevtoolsControlMsg>>

A channel for the constellation to send messages to the devtools thread.

bluetooth_thread: IpcSender<BluetoothRequest>

An IPC channel for the constellation to send messages to the bluetooth thread.

swmanager_chan: Option<IpcSender<ServiceWorkerMsg>>

An IPC channel for the constellation to send messages to the Service Worker Manager thread.

swmanager_sender: IpcSender<SWManagerMsg>

An IPC channel for Service Worker Manager threads to send messages to the constellation. This is the SW Manager thread's view of swmanager_receiver.

swmanager_receiver: Receiver<Result<SWManagerMsg, IpcError>>

A channel for the constellation to receive messages from the Service Worker Manager thread. This is the constellation's view of swmanager_sender.

time_profiler_chan: ProfilerChan

A channel for the constellation to send messages to the time profiler thread.

mem_profiler_chan: ProfilerChan

A channel for the constellation to send messages to the memory profiler thread.

scheduler_chan: IpcSender<TimerEventRequest>

A channel for the constellation to send messages to the timer thread.

webrender_api_sender: RenderApiSender

A channel for the constellation to send messages to the Webrender thread.

event_loops: HashMap<FrameId, HashMap<String, Weak<EventLoop>>>

The set of all event loops in the browser. We generate a new event loop for each registered domain name (aka eTLD+1) in each top-level frame. We store the event loops in a map indexed by top-level frame id (as a FrameId) and registered domain name (as a String) to event loops. This double indirection ensures that separate tabs do not share event loops, even if the same domain is loaded in each. It is important that scripts with the same eTLD+1 share an event loop, since they can use document.domain to become same-origin, at which point they can share DOM objects.

pipelines: HashMap<PipelineId, Pipeline>

The set of all the pipelines in the browser. (See the pipeline module for more details.)

frames: HashMap<FrameId, Frame>

The set of all the frames in the browser.

pending_frames: Vec<FrameChange>

When a navigation is performed, we do not immediately update the frame tree, instead we ask the event loop to begin loading the new document, and do not update the frame tree until the document is active. Between starting the load and it activating, we store a FrameChange object for the navigation in progress.

root_frame_id: FrameId

The root frame.

focus_pipeline_id: Option<PipelineId>

The currently focused pipeline for key events.

next_pipeline_namespace_id: PipelineNamespaceId

Pipeline IDs are namespaced in order to avoid name collisions, and the namespaces are allocated by the constellation.

window_size: WindowSizeData

The size of the top-level window.

webdriver: WebDriverData

Bits of state used to interact with the webdriver implementation

document_states: HashMap<PipelineId, DocumentState>

Document states for loaded pipelines (used only when writing screenshots).

shutting_down: bool

Are we shutting down?

handled_warnings: VecDeque<(Option<String>, String)>

Have we seen any warnings? Hopefully always empty! The buffer contains (thread_name, reason) entries.

random_pipeline_closure: Option<(ServoRng, f32)>

The random number generator and probability for closing pipelines. This is for testing the hardening of the constellation.

phantom: PhantomData<(Message, LTF, STF)>

Phantom data that keeps the Rust type system happy.

webvr_thread: Option<IpcSender<WebVRMsg>>

A channel through which messages can be sent to the webvr thread.

Methods

impl<Message, LTF, STF> Constellation<Message, LTF, STF> where LTF: LayoutThreadFactory<Message=Message>,
STF: ScriptThreadFactory<Message=Message>

fn start(state: InitialConstellationState)
-> (Sender<FromCompositorMsg>, IpcSender<SWManagerMsg>)

Create a new constellation thread.

fn run(&mut self)

The main event loop for the constellation.

fn next_pipeline_namespace_id(&mut self) -> PipelineNamespaceId

Generate a new pipeline id namespace.

fn new_pipeline(&mut self,
pipeline_id: PipelineId,
frame_id: FrameId,
parent_info: Option<(PipelineId, FrameType)>,
initial_window_size: Option<TypedSize2D<f32, CSSPixel>>,
load_data: LoadData,
sandbox: IFrameSandboxState,
is_private: bool)

Helper function for creating a pipeline

fn current_frame_tree_iter(&self, frame_id_root: FrameId) -> FrameTreeIterator

Get an iterator for the current frame tree. Specify self.root_frame_id to iterate the entire tree, or a specific frame id to iterate only that sub-tree. Iterates over the fully active frames in the tree.

fn full_frame_tree_iter(&self, frame_id_root: FrameId) -> FullFrameTreeIterator

Get an iterator for the current frame tree. Specify self.root_frame_id to iterate the entire tree, or a specific frame id to iterate only that sub-tree. Iterates over all frames in the tree.

fn joint_session_future<'a>(&'a self,
frame_id_root: FrameId)
-> impl Iterator<Item=FrameState>

The joint session future is the merge of the session future of every frame in the frame tree, sorted chronologically.

fn joint_session_future_is_empty(&self, frame_id_root: FrameId) -> bool

Is the joint session future empty?

fn joint_session_past<'a>(&self,
frame_id_root: FrameId)
-> impl Iterator<Item=FrameState>

The joint session past is the merge of the session past of every frame in the frame tree, sorted reverse chronologically.

fn joint_session_past_is_empty(&self, frame_id_root: FrameId) -> bool

Is the joint session past empty?

fn new_frame(&mut self,
frame_id: FrameId,
pipeline_id: PipelineId,
url: ServoUrl)

Create a new frame and update the internal bookkeeping.

fn handle_request(&mut self)

Handles loading pages, navigation, and granting access to the compositor

fn handle_request_from_swmanager(&mut self, message: SWManagerMsg)

fn handle_request_from_compositor(&mut self, message: FromCompositorMsg)

fn handle_request_from_script(&mut self, message: FromScriptMsg)

fn handle_request_from_layout(&mut self, message: FromLayoutMsg)

fn handle_register_serviceworker(&self,
scope_things: ScopeThings,
scope: ServoUrl)

fn handle_broadcast_storage_event(&self,
pipeline_id: PipelineId,
storage: StorageType,
url: ServoUrl,
key: Option<String>,
old_value: Option<String>,
new_value: Option<String>)

fn handle_exit(&mut self)

fn handle_shutdown(&mut self)

fn handle_pipeline_exited(&mut self, pipeline_id: PipelineId)

fn handle_send_error(&mut self, pipeline_id: PipelineId, err: IpcError)

fn handle_panic(&mut self,
top_level_frame_id: FrameId,
reason: String,
backtrace: Option<String>)

fn handle_log_entry(&mut self,
top_level_frame_id: Option<FrameId>,
thread_name: Option<String>,
entry: LogEntry)

fn handle_webvr_event(&mut self, ids: Vec<PipelineId>, event: WebVREventMsg)

fn handle_init_load(&mut self, url: ServoUrl)

fn handle_frame_size_msg(&mut self,
iframe_sizes: Vec<(PipelineId, TypedSize2D<f32, CSSPixel>)>)

fn handle_subframe_loaded(&mut self, pipeline_id: PipelineId)

fn handle_script_loaded_url_in_iframe_msg(&mut self,
load_info: IFrameLoadInfoWithData)

fn handle_script_loaded_about_blank_in_iframe_msg(&mut self,
load_info: IFrameLoadInfo,
layout_sender: IpcSender<LayoutControlMsg>)

fn handle_set_cursor_msg(&mut self, cursor: Cursor)

fn handle_change_running_animations_state(&mut self,
pipeline_id: PipelineId,
animation_state: AnimationState)

fn handle_tick_animation(&mut self,
pipeline_id: PipelineId,
tick_type: AnimationTickType)

fn handle_alert(&mut self,
pipeline_id: PipelineId,
message: String,
sender: IpcSender<bool>)

fn handle_load_url_msg(&mut self,
source_id: PipelineId,
load_data: LoadData,
replace: bool)

fn load_url(&mut self,
source_id: PipelineId,
load_data: LoadData,
replace: bool)
-> Option<PipelineId>

fn handle_load_start_msg(&mut self, pipeline_id: PipelineId)

fn handle_load_complete_msg(&mut self, pipeline_id: PipelineId)

fn handle_traverse_history_msg(&mut self,
pipeline_id: Option<PipelineId>,
direction: TraversalDirection)

fn handle_joint_session_history_length(&self,
pipeline_id: PipelineId,
sender: IpcSender<u32>)

fn handle_key_msg(&mut self,
ch: Option<char>,
key: Key,
state: KeyState,
mods: KeyModifiers)

fn handle_reload_msg(&mut self)

fn handle_get_pipeline_title_msg(&mut self, pipeline_id: PipelineId)

fn handle_mozbrowser_event_msg(&mut self,
parent_pipeline_id: PipelineId,
pipeline_id: PipelineId,
event: MozBrowserEvent)

fn handle_get_pipeline(&mut self,
frame_id: Option<FrameId>,
resp_chan: IpcSender<Option<PipelineId>>)

fn handle_get_frame(&mut self,
pipeline_id: PipelineId,
resp_chan: IpcSender<Option<FrameId>>)

fn focus_parent_pipeline(&mut self, pipeline_id: PipelineId)

fn handle_focus_msg(&mut self, pipeline_id: PipelineId)

fn handle_remove_iframe_msg(&mut self, frame_id: FrameId) -> Vec<PipelineId>

fn handle_set_visible_msg(&mut self, pipeline_id: PipelineId, visible: bool)

fn handle_visibility_change_complete(&mut self,
pipeline_id: PipelineId,
visibility: bool)

fn handle_create_canvas_paint_thread_msg(&mut self,
size: &Size2D<i32>,
response_sender: IpcSender<IpcSender<CanvasMsg>>)

fn handle_create_webgl_paint_thread_msg(&mut self,
size: &Size2D<i32>,
attributes: GLContextAttributes,
response_sender: IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>)

fn handle_webdriver_msg(&mut self, msg: WebDriverCommandMsg)

fn traverse_to_entry(&mut self, entry: FrameState)

fn get_top_level_frame_for_pipeline(&self, pipeline_id: PipelineId) -> FrameId

fn load_url_for_webdriver(&mut self,
pipeline_id: PipelineId,
load_data: LoadData,
reply: IpcSender<LoadStatus>,
replace: bool)

fn add_or_replace_pipeline_in_frame_tree(&mut self, frame_change: FrameChange)

fn handle_activate_document_msg(&mut self, pipeline_id: PipelineId)

fn handle_window_size_msg(&mut self,
new_size: WindowSizeData,
size_type: WindowSizeType)

Called when the window is resized.

fn handle_viewport_constrained_msg(&mut self,
pipeline_id: PipelineId,
constraints: ViewportConstraints)

Handle updating actual viewport / zoom due to @viewport rules

fn handle_is_ready_to_save_image(&mut self,
pipeline_states: HashMap<PipelineId, Epoch>)
-> ReadyToSave

Checks the state of all script and layout pipelines to see if they are idle and compares the current layout state to what the compositor has. This is used to check if the output image is "stable" and can be written as a screenshot for reftests. Since this function is only used in reftests, we do not harden it against panic.

fn get_activity(&self, pipeline_id: PipelineId) -> DocumentActivity

Get the current activity of a pipeline.

fn set_activity(&self, pipeline_id: PipelineId, activity: DocumentActivity)

Set the current activity of a pipeline.

fn update_activity(&self, pipeline_id: PipelineId)

Update the current activity of a pipeline.

fn clear_joint_session_future(&mut self, frame_id: FrameId)

fn close_frame(&mut self, frame_id: FrameId, exit_mode: ExitPipelineMode)

fn close_frame_children(&mut self,
frame_id: FrameId,
dbc: DiscardBrowsingContext,
exit_mode: ExitPipelineMode)

fn close_pipeline(&mut self,
pipeline_id: PipelineId,
dbc: DiscardBrowsingContext,
exit_mode: ExitPipelineMode)

fn maybe_close_random_pipeline(&mut self)

fn frame_to_sendable(&self, frame_id: FrameId) -> Option<SendableFrameTree>

fn send_frame_tree(&mut self)

fn trigger_mozbrowserlocationchange(&self, pipeline_id: PipelineId)

fn trigger_mozbrowsererror(&mut self,
top_level_frame_id: FrameId,
reason: String,
backtrace: Option<String>)

fn focused_pipeline_in_tree(&self, frame_id: FrameId) -> bool

fn pipeline_is_in_current_frame(&self, pipeline_id: PipelineId) -> bool

fn pipeline_exists_in_tree(&self,
pipeline_id: PipelineId,
root_frame_id: FrameId)
-> bool