Struct constellation::webview::WebViewManager

pub struct WebViewManager<WebView> {
    webviews: HashMap<TopLevelBrowsingContextId, WebView>,
    focus_order: Vec<TopLevelBrowsingContextId>,
    is_focused: bool,
}

Fields
webviews: HashMap<TopLevelBrowsingContextId, WebView>

Our top-level browsing contexts. In the WebRender scene, their pipelines are the children of a single root pipeline that also applies any pinch zoom transformation.

focus_order: Vec<TopLevelBrowsingContextId>

The order in which they were focused, latest last.

is_focused: bool

Whether the latest webview in focus order is currently focused.

Implementations

impl<WebView> WebViewManager<WebView>

pub fn add( &mut self, top_level_browsing_context_id: TopLevelBrowsingContextId, webview: WebView )

pub fn remove( &mut self, top_level_browsing_context_id: TopLevelBrowsingContextId ) -> Option<WebView>

pub fn get( &self, top_level_browsing_context_id: TopLevelBrowsingContextId ) -> Option<&WebView>

pub fn get_mut( &mut self, top_level_browsing_context_id: TopLevelBrowsingContextId ) -> Option<&mut WebView>

pub fn focused_webview(&self) -> Option<(TopLevelBrowsingContextId, &WebView)>

pub fn focus( &mut self, top_level_browsing_context_id: TopLevelBrowsingContextId )

pub fn unfocus(&mut self)