Module constellation::browsingcontext

Structs

AllBrowsingContextsIterator

An iterator over browsing contexts, returning all descendant contexts in depth-first order. Note that this iterator returns all contexts, not just the fully active ones.

BrowsingContext

The constellation’s view of a browsing context. Each browsing context has a session history, caused by navigation and traversing the history. Each browsing context has its current entry, plus past and future entries. The past is sorted chronologically, the future is sorted reverse chronologically: in particular prev.pop() is the latest past entry, and next.pop() is the earliest future entry.

FullyActiveBrowsingContextsIterator

An iterator over browsing contexts, returning the descendant contexts whose active documents are fully active, in depth-first order.

NewBrowsingContextInfo

Because a browsing context is only constructed once the document that’s going to be in it becomes active (i.e. not when a pipeline is spawned), some values needed in browsing context are not easily available at the point of constructing it. Thus, every time a pipeline is created for a browsing context which doesn’t exist yet, these values needed for the new browsing context are stored here so that they may be available later.