Module constellation::constellation

The Constellation, Servo's Grand Central Station

The primary duty of a Constellation is to mediate between the graphics compositor and the many Pipelines in the browser's navigation context, each Pipeline encompassing a ScriptThread, LayoutThread, and PaintThread.

Structs

Constellation
Maintains the pipelines and navigation context and grants permission to composite.

Frame
Stores the navigation context for a single frame in the frame tree.

FrameChange
Represents a pending change in the frame tree, that will be applied once the new pipeline has loaded and completed initial layout / paint.

FrameTreeIterator
An iterator over a frame tree, returning nodes in depth-first order. Note that this iterator should not be used to mutate nodes during iteration. Mutating nodes once the iterator is out of scope is OK.

InitialConstellationState
State needed to construct a constellation.

WebDriverData

Enums

ChildProcess

ExitPipelineMode

ReadyToSave