Struct constellation::constellation::Constellation

pub struct Constellation<Message, LTF, STF> {
    script_sender: IpcSender<FromScriptMsg>,
    layout_sender: IpcSender<FromLayoutMsg>,
    script_receiver: Receiver<FromScriptMsg>,
    compositor_receiver: Receiver<FromCompositorMsg>,
    layout_receiver: Receiver<FromLayoutMsg>,
    compositor_proxy: Box<CompositorProxy>,
    public_resource_threads: ResourceThreads,
    private_resource_threads: ResourceThreads,
    image_cache_thread: ImageCacheThread,
    debugger_chan: Option<Sender>,
    devtools_chan: Option<Sender<DevtoolsControlMsg>>,
    bluetooth_thread: IpcSender<BluetoothRequest>,
    swmanager_chan: Option<IpcSender<ServiceWorkerMsg>>,
    swmanager_sender: IpcSender<SWManagerMsg>,
    swmanager_receiver: Receiver<SWManagerMsg>,
    pipelines: HashMap<PipelineId, Pipeline>,
    frames: HashMap<FrameId, Frame>,
    font_cache_thread: FontCacheThread,
    root_frame_id: FrameId,
    next_pipeline_namespace_id: PipelineNamespaceId,
    focus_pipeline_id: Option<PipelineId>,
    pending_frames: Vec<FrameChange>,
    time_profiler_chan: ProfilerChan,
    mem_profiler_chan: ProfilerChan,
    phantom: PhantomData<(Message, LTF, STF)>,
    window_size: WindowSizeData,
    webdriver: WebDriverData,
    scheduler_chan: IpcSender<TimerEventRequest>,
    child_processes: Vec<ChildProcess>,
    document_states: HashMap<PipelineId, DocumentState>,
    webrender_api_sender: RenderApiSender,
    shutting_down: bool,
    handled_warnings: VecDeque<(Option<String>, String)>,
    random_pipeline_closure: Option<(StdRng, f32)>,
}

Maintains the pipelines and navigation context and grants permission to composite.

It is parameterized over a LayoutThreadFactory and a ScriptThreadFactory (which in practice are implemented by LayoutThread in the layout crate, and ScriptThread in the script crate).

Fields
script_sender: IpcSender<FromScriptMsg>

A channel through which script messages can be sent to this object.

layout_sender: IpcSender<FromLayoutMsg>

A channel through which layout thread messages can be sent to this object.

script_receiver: Receiver<FromScriptMsg>

Receives messages from scripts.

compositor_receiver: Receiver<FromCompositorMsg>

Receives messages from the compositor

layout_receiver: Receiver<FromLayoutMsg>

Receives messages from the layout thread

compositor_proxy: Box<CompositorProxy>

A channel (the implementation of which is port-specific) through which messages can be sent to the compositor.

public_resource_threads: ResourceThreads

Channels through which messages can be sent to the resource-related threads.

private_resource_threads: ResourceThreads

Channels through which messages can be sent to the resource-related threads.

image_cache_thread: ImageCacheThread

A channel through which messages can be sent to the image cache thread.

debugger_chan: Option<Sender>

A channel through which messages can be sent to the debugger.

devtools_chan: Option<Sender<DevtoolsControlMsg>>

A channel through which messages can be sent to the developer tools.

bluetooth_thread: IpcSender<BluetoothRequest>

A channel through which messages can be sent to the bluetooth thread.

swmanager_chan: Option<IpcSender<ServiceWorkerMsg>>

Sender to Service Worker Manager thread

swmanager_sender: IpcSender<SWManagerMsg>

to send messages to this object

swmanager_receiver: Receiver<SWManagerMsg>

to receive sw manager message

pipelines: HashMap<PipelineId, Pipeline>

A list of all the pipelines. (See the pipeline module for more details.)

frames: HashMap<FrameId, Frame>

A list of all the frames

font_cache_thread: FontCacheThread

A channel through which messages can be sent to the font cache.

root_frame_id: FrameId

ID of the root frame.

next_pipeline_namespace_id: PipelineNamespaceId

The next free ID to assign to a pipeline ID namespace.

focus_pipeline_id: Option<PipelineId>

Pipeline ID that has currently focused element for key events.

pending_frames: Vec<FrameChange>

Navigation operations that are in progress.

time_profiler_chan: ProfilerChan

A channel through which messages can be sent to the time profiler.

mem_profiler_chan: ProfilerChan

A channel through which messages can be sent to the memory profiler.

phantom: PhantomData<(Message, LTF, STF)> window_size: WindowSizeData webdriver: WebDriverData

Bits of state used to interact with the webdriver implementation

scheduler_chan: IpcSender<TimerEventRequest> child_processes: Vec<ChildProcess>

A list of child content processes.

document_states: HashMap<PipelineId, DocumentState>

Document states for loaded pipelines (used only when writing screenshots).

webrender_api_sender: RenderApiSender shutting_down: bool

Are we shutting down?

handled_warnings: VecDeque<(Option<String>, String)>

Have we seen any warnings? Hopefully always empty! The buffer contains (thread_name, reason) entries.

random_pipeline_closure: Option<(StdRng, f32)>

The random number generator and probability for closing pipelines. This is for testing the hardening of the constellation.

Methods

impl<Message, LTF, STF> Constellation<Message, LTF, STF> where LTF: LayoutThreadFactory<Message=Message>, STF: ScriptThreadFactory<Message=Message>

fn start(state: InitialConstellationState)
-> (Sender<FromCompositorMsg>, IpcSender<SWManagerMsg>)

fn run(&mut self)

fn next_pipeline_namespace_id(&mut self) -> PipelineNamespaceId

fn new_pipeline(&mut self,
pipeline_id: PipelineId,
frame_id: FrameId,
parent_info: Option<(PipelineId, FrameType)>,
old_pipeline_id: Option<PipelineId>,
initial_window_size: Option<TypedSize2D<f32, PagePx>>,
script_channel: Option<IpcSender<ConstellationControlMsg>>,
load_data: LoadData,
is_private: bool)

Helper function for creating a pipeline

fn current_frame_tree_iter(&self, frame_id_root: FrameId) -> FrameTreeIterator

fn full_frame_tree_iter(&self, frame_id_root: FrameId) -> FullFrameTreeIterator

fn joint_session_future(&self,
frame_id_root: FrameId)
-> Vec<(Instant, FrameId, PipelineId)>

fn joint_session_future_is_empty(&self, frame_id_root: FrameId) -> bool

fn joint_session_past(&self,
frame_id_root: FrameId)
-> Vec<(Instant, FrameId, PipelineId)>

fn joint_session_past_is_empty(&self, frame_id_root: FrameId) -> bool

fn new_frame(&mut self, frame_id: FrameId, pipeline_id: PipelineId)

fn handle_request(&mut self)

Handles loading pages, navigation, and granting access to the compositor

fn handle_request_from_swmanager(&mut self, message: SWManagerMsg)

fn handle_request_from_compositor(&mut self, message: FromCompositorMsg)

fn handle_request_from_script(&mut self, message: FromScriptMsg)

fn handle_request_from_layout(&mut self, message: FromLayoutMsg)

fn handle_register_serviceworker(&self, scope_things: ScopeThings, scope: Url)

fn handle_exit(&mut self)

fn handle_shutdown(&mut self)

fn handle_pipeline_exited(&mut self, pipeline_id: PipelineId)

fn handle_send_error(&mut self, pipeline_id: PipelineId, err: IOError)

fn handle_panic(&mut self,
pipeline_id: Option<PipelineId>,
reason: String,
backtrace: Option<String>)

fn handle_log_entry(&mut self,
pipeline_id: Option<PipelineId>,
thread_name: Option<String>,
entry: LogEntry)

fn handle_init_load(&mut self, url: Url)

fn handle_frame_size_msg(&mut self,
pipeline_id: PipelineId,
size: &TypedSize2D<f32, PagePx>)

fn handle_subframe_loaded(&mut self, pipeline_id: PipelineId)

fn handle_script_loaded_url_in_iframe_msg(&mut self, load_info: IFrameLoadInfo)

fn handle_set_cursor_msg(&mut self, cursor: Cursor)

fn handle_change_running_animations_state(&mut self,
pipeline_id: PipelineId,
animation_state: AnimationState)

fn handle_tick_animation(&mut self,
pipeline_id: PipelineId,
tick_type: AnimationTickType)

fn handle_alert(&mut self,
pipeline_id: PipelineId,
message: String,
sender: IpcSender<bool>)

fn handle_load_url_msg(&mut self,
source_id: PipelineId,
load_data: LoadData,
replace: bool)

fn load_url(&mut self,
source_id: PipelineId,
load_data: LoadData,
replace: bool)
-> Option<PipelineId>

fn handle_load_start_msg(&mut self, pipeline_id: PipelineId)

fn handle_load_complete_msg(&mut self, pipeline_id: PipelineId)

fn handle_traverse_history_msg(&mut self,
pipeline_id: Option<PipelineId>,
direction: TraversalDirection)

fn handle_joint_session_history_length(&self,
pipeline_id: PipelineId,
sender: IpcSender<u32>)

fn handle_key_msg(&mut self,
ch: Option<char>,
key: Key,
state: KeyState,
mods: KeyModifiers)

fn handle_reload_msg(&mut self)

fn handle_get_pipeline_title_msg(&mut self, pipeline_id: PipelineId)

fn handle_mozbrowser_event_msg(&mut self,
parent_pipeline_id: PipelineId,
pipeline_id: PipelineId,
event: MozBrowserEvent)

fn handle_get_pipeline(&mut self,
frame_id: Option<FrameId>,
resp_chan: IpcSender<Option<(PipelineId, bool)>>)

fn handle_get_frame(&mut self,
pipeline_id: PipelineId,
resp_chan: IpcSender<Option<FrameId>>)

fn focus_parent_pipeline(&mut self, pipeline_id: PipelineId)

fn handle_focus_msg(&mut self, pipeline_id: PipelineId)

fn handle_remove_iframe_msg(&mut self, pipeline_id: PipelineId)

fn handle_set_visible_msg(&mut self, pipeline_id: PipelineId, visible: bool)

fn handle_visibility_change_complete(&mut self,
pipeline_id: PipelineId,
visibility: bool)

fn handle_create_canvas_paint_thread_msg(&mut self,
size: &Size2D<i32>,
response_sender: IpcSender<IpcSender<CanvasMsg>>)

fn handle_create_webgl_paint_thread_msg(&mut self,
size: &Size2D<i32>,
attributes: GLContextAttributes,
response_sender: IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>)

fn handle_webdriver_msg(&mut self, msg: WebDriverCommandMsg)

fn traverse_frame_to_pipeline(&mut self,
frame_id: FrameId,
next_pipeline_id: PipelineId)

fn get_top_level_frame_for_pipeline(&self,
pipeline_id: Option<PipelineId>)
-> FrameId

fn load_url_for_webdriver(&mut self,
pipeline_id: PipelineId,
load_data: LoadData,
reply: IpcSender<LoadStatus>,
replace: bool)

fn add_or_replace_pipeline_in_frame_tree(&mut self, frame_change: FrameChange)

fn handle_activate_document_msg(&mut self, pipeline_id: PipelineId)

fn handle_window_size_msg(&mut self,
new_size: WindowSizeData,
size_type: WindowSizeType)

Called when the window is resized.

fn handle_viewport_constrained_msg(&mut self,
pipeline_id: PipelineId,
constraints: ViewportConstraints)

Handle updating actual viewport / zoom due to @viewport rules

fn handle_is_ready_to_save_image(&mut self,
pipeline_states: HashMap<PipelineId, Epoch>)
-> ReadyToSave

Checks the state of all script and layout pipelines to see if they are idle and compares the current layout state to what the compositor has. This is used to check if the output image is "stable" and can be written as a screenshot for reftests. Since this function is only used in reftests, we do not harden it against panic.

fn clear_joint_session_future(&mut self, frame_id: FrameId)

fn close_frame(&mut self, frame_id: FrameId, exit_mode: ExitPipelineMode)

fn close_pipeline(&mut self,
pipeline_id: PipelineId,
exit_mode: ExitPipelineMode)

fn maybe_close_random_pipeline(&mut self)

fn frame_to_sendable(&self, frame_id: FrameId) -> Option<SendableFrameTree>

fn send_frame_tree(&mut self)

fn get_mozbrowser_ancestor_info(&self,
original_pipeline_id: PipelineId)
-> Option<(PipelineId, FrameId)>

For a given pipeline, determine the mozbrowser iframe that transitively contains it. There could be arbitrary levels of nested iframes in between them.

fn trigger_mozbrowserlocationchange(&self, pipeline_id: PipelineId)

fn trigger_mozbrowsererror(&mut self,
pipeline_id: Option<PipelineId>,
reason: String,
backtrace: Option<String>)

fn focused_pipeline_in_tree(&self, frame_id: FrameId) -> bool

fn pipeline_is_in_current_frame(&self, pipeline_id: PipelineId) -> bool

fn pipeline_exists_in_tree(&self,
pipeline_id: PipelineId,
root_frame_id: FrameId)
-> bool