Struct constellation::frame::FrameChange

pub struct FrameChange {
    pub frame_id: FrameId,
    pub new_pipeline_id: PipelineId,
    pub load_data: LoadData,
    pub replace_instant: Option<Instant>,
}

Represents a pending change in the frame tree, that will be applied once the new pipeline has loaded and completed initial layout / paint.

Fields
frame_id: FrameId

The frame to change.

new_pipeline_id: PipelineId

The pipeline for the document being loaded.

load_data: LoadData

The data for the document being loaded.

replace_instant: Option<Instant>

Is the new document replacing the current document (e.g. a reload) or pushing it into the session history (e.g. a navigation)? If it is replacing an existing entry, we store its timestamp.