Struct constellation::pipeline::InitialPipelineState

pub struct InitialPipelineState {
    pub id: PipelineId,
    pub frame_id: FrameId,
    pub top_level_frame_id: FrameId,
    pub parent_info: Option<(PipelineId, FrameType)>,
    pub constellation_chan: IpcSender<ScriptMsg>,
    pub layout_to_constellation_chan: IpcSender<LayoutMsg>,
    pub scheduler_chan: IpcSender<TimerEventRequest>,
    pub compositor_proxy: Box<CompositorProxy + Send + 'static>,
    pub devtools_chan: Option<Sender<DevtoolsControlMsg>>,
    pub bluetooth_thread: IpcSender<BluetoothRequest>,
    pub swmanager_thread: IpcSender<SWManagerMsg>,
    pub image_cache_thread: ImageCacheThread,
    pub font_cache_thread: FontCacheThread,
    pub resource_threads: ResourceThreads,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub window_size: Option<TypedSize2D<f32, CSSPixel>>,
    pub device_pixel_ratio: ScaleFactor<f32, CSSPixel, DevicePixel>,
    pub event_loop: Option<Rc<EventLoop>>,
    pub load_data: LoadData,
    pub pipeline_namespace_id: PipelineNamespaceId,
    pub prev_visibility: Option<bool>,
    pub webrender_api_sender: RenderApiSender,
    pub is_private: bool,
    pub webvr_thread: Option<IpcSender<WebVRMsg>>,
}

Initial setup data needed to construct a pipeline.

DO NOT add any Senders to this unless you absolutely know what you're doing, or pcwalton will have to rewrite your code. Use IPC senders instead.

Fields
id: PipelineId

The ID of the pipeline to create.

frame_id: FrameId

The ID of the frame that contains this Pipeline.

top_level_frame_id: FrameId

The ID of the top-level frame that contains this Pipeline.

parent_info: Option<(PipelineId, FrameType)>

The ID of the parent pipeline and frame type, if any. If None, this is the root.

constellation_chan: IpcSender<ScriptMsg>

A channel to the associated constellation.

layout_to_constellation_chan: IpcSender<LayoutMsg>

A channel for the layout thread to send messages to the constellation.

scheduler_chan: IpcSender<TimerEventRequest>

A channel to schedule timer events.

compositor_proxy: Box<CompositorProxy + Send + 'static>

A channel to the compositor.

devtools_chan: Option<Sender<DevtoolsControlMsg>>

A channel to the developer tools, if applicable.

bluetooth_thread: IpcSender<BluetoothRequest>

A channel to the bluetooth thread.

swmanager_thread: IpcSender<SWManagerMsg>

A channel to the service worker manager thread

image_cache_thread: ImageCacheThread

A channel to the image cache thread.

font_cache_thread: FontCacheThread

A channel to the font cache thread.

resource_threads: ResourceThreads

Channels to the resource-related threads.

time_profiler_chan: ProfilerChan

A channel to the time profiler thread.

mem_profiler_chan: ProfilerChan

A channel to the memory profiler thread.

window_size: Option<TypedSize2D<f32, CSSPixel>>

Information about the initial window size.

device_pixel_ratio: ScaleFactor<f32, CSSPixel, DevicePixel>

Information about the device pixel ratio.

event_loop: Option<Rc<EventLoop>>

The event loop to run in, if applicable.

load_data: LoadData

Information about the page to load.

pipeline_namespace_id: PipelineNamespaceId

The ID of the pipeline namespace for this script thread.

prev_visibility: Option<bool>

Pipeline visibility to be inherited

webrender_api_sender: RenderApiSender

Webrender api.

is_private: bool

Whether this pipeline is considered private.

webvr_thread: Option<IpcSender<WebVRMsg>>

A channel to the webvr thread.