Struct constellation::pipeline::InitialPipelineState

pub struct InitialPipelineState {
    pub id: PipelineId,
    pub parent_info: Option<(PipelineId, SubpageId, FrameType)>,
    pub constellation_chan: IpcSender<ScriptMsg>,
    pub layout_to_constellation_chan: IpcSender<LayoutMsg>,
    pub panic_chan: IpcSender<PanicMsg>,
    pub scheduler_chan: IpcSender<TimerEventRequest>,
    pub compositor_proxy: Box<CompositorProxy + 'static + Send>,
    pub devtools_chan: Option<Sender<DevtoolsControlMsg>>,
    pub bluetooth_thread: IpcSender<BluetoothMethodMsg>,
    pub image_cache_thread: ImageCacheThread,
    pub font_cache_thread: FontCacheThread,
    pub resource_threads: ResourceThreads,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub window_size: Option<TypedSize2D<PagePx, f32>>,
    pub device_pixel_ratio: ScaleFactor<ViewportPx, DevicePixel, f32>,
    pub script_chan: Option<IpcSender<ConstellationControlMsg>>,
    pub load_data: LoadData,
    pub pipeline_namespace_id: PipelineNamespaceId,
    pub webrender_api_sender: Option<RenderApiSender>,
}

Initial setup data needed to construct a pipeline.

DO NOT add any Senders to this unless you absolutely know what you're doing, or pcwalton will have to rewrite your code. Use IPC senders instead.

Fields

id

The ID of the pipeline to create.

parent_info

The subpage ID of this pipeline to create in its pipeline parent. If None, this is the root.

constellation_chan

A channel to the associated constellation.

layout_to_constellation_chan

A channel for the layout thread to send messages to the constellation.

panic_chan

A channel to report panics

scheduler_chan

A channel to schedule timer events.

compositor_proxy

A channel to the compositor.

devtools_chan

A channel to the developer tools, if applicable.

bluetooth_thread

A channel to the bluetooth thread.

image_cache_thread

A channel to the image cache thread.

font_cache_thread

A channel to the font cache thread.

resource_threads

Channels to the resource-related threads.

time_profiler_chan

A channel to the time profiler thread.

mem_profiler_chan

A channel to the memory profiler thread.

window_size

Information about the initial window size.

device_pixel_ratio

Information about the device pixel ratio.

script_chan

A channel to the script thread, if applicable. If this is Some, then parent_info must also be Some.

load_data

Information about the page to load.

pipeline_namespace_id

The ID of the pipeline namespace for this script thread.

webrender_api_sender

Optional webrender api (if enabled).