Struct constellation::pipeline::Pipeline

pub struct Pipeline {
    pub id: PipelineId,
    pub frame_id: FrameId,
    pub parent_info: Option<(PipelineId, FrameType)>,
    pub event_loop: Rc<EventLoop>,
    pub layout_chan: IpcSender<LayoutControlMsg>,
    pub compositor_proxy: Box<CompositorProxy + Send + 'static>,
    pub url: ServoUrl,
    pub title: Option<String>,
    pub size: Option<TypedSize2D<f32, PagePx>>,
    pub running_animations: bool,
    pub children: Vec<FrameId>,
    pub is_private: bool,
    pub visible: bool,
}

A Pipeline is the constellation's view of a Document. Each pipeline has an event loop (executed by a script thread) and a layout thread. A script thread may be responsible for many pipelines, but a layout thread is only responsible for one.

Fields
id: PipelineId

The ID of the pipeline.

frame_id: FrameId

The ID of the frame that contains this Pipeline.

parent_info: Option<(PipelineId, FrameType)>

The parent pipeline of this one. None if this is a root pipeline. Note that because of mozbrowser iframes, even top-level pipelines may have a parent (in which case the frame type will be MozbrowserIFrame). TODO: move this field to Frame.

event_loop: Rc<EventLoop>

The event loop handling this pipeline.

layout_chan: IpcSender<LayoutControlMsg>

A channel to layout, for performing reflows and shutdown.

compositor_proxy: Box<CompositorProxy + Send + 'static>

A channel to the compositor.

url: ServoUrl

The most recently loaded URL in this pipeline. Note that this URL can change, for example if the page navigates to a hash URL.

title: Option<String>

The title of the most recently-loaded page.

size: Option<TypedSize2D<f32, PagePx>>

The size of the frame. TODO: move this field to Frame.

running_animations: bool

Whether this pipeline is currently running animations. Pipelines that are running animations cause composites to be continually scheduled.

children: Vec<FrameId>

The child frames of this pipeline (these are iframes in the document).

is_private: bool

Whether this pipeline is in private browsing mode. TODO: move this field to Frame.

visible: bool

Whether this pipeline should be treated as visible for the purposes of scheduling and resource management.

Methods

impl Pipeline

fn spawn<Message, LTF, STF>(state: InitialPipelineState)
-> Result<Pipeline, Error> where LTF: LayoutThreadFactory<Message=Message>,
STF: ScriptThreadFactory<Message=Message>

Starts a layout thread, and possibly a script thread, in a new process if requested.

fn new(id: PipelineId,
frame_id: FrameId,
parent_info: Option<(PipelineId, FrameType)>,
event_loop: Rc<EventLoop>,
layout_chan: IpcSender<LayoutControlMsg>,
compositor_proxy: Box<CompositorProxy + Send + 'static>,
is_private: bool,
url: ServoUrl,
size: Option<TypedSize2D<f32, PagePx>>,
visible: bool)
-> Pipeline

Creates a new Pipeline, after the script and layout threads have been spawned.

fn exit(&self, discard_bc: DiscardBrowsingContext)

A normal exit of the pipeline, which waits for the compositor, and delegates layout shutdown to the script thread.

fn force_exit(&self, discard_bc: DiscardBrowsingContext)

A forced exit of the shutdown, which does not wait for the compositor, or for the script thread to shut down layout.

fn set_activity(&self, activity: DocumentActivity)

Notify this pipeline of its activity.

fn to_sendable(&self) -> CompositionPipeline

The compositor's view of a pipeline.

fn add_child(&mut self, frame_id: FrameId)

Add a new child frame.

fn remove_child(&mut self, frame_id: FrameId)

Remove a child frame.

fn trigger_mozbrowser_event(&self,
child_id: Option<FrameId>,
event: MozBrowserEvent)

Send a mozbrowser event to the script thread for this pipeline. This will cause an event to be fired on an iframe in the document, or on the Window if no frame is given.

fn notify_visibility(&self)

Notify the script thread that this pipeline is visible.

fn change_visibility(&mut self, visible: bool)

Change the visibility of this pipeline.