Struct constellation::pipeline::Pipeline

pub struct Pipeline {
    pub id: PipelineId,
    pub parent_info: Option<(PipelineId, SubpageId, FrameType)>,
    pub script_chan: IpcSender<ConstellationControlMsg>,
    pub layout_chan: IpcSender<LayoutControlMsg>,
    pub compositor_proxy: Box<CompositorProxy + 'static + Send>,
    pub chrome_to_paint_chan: Sender<ChromeToPaintMsg>,
    pub url: Url,
    pub title: Option<String>,
    pub size: Option<TypedSize2D<PagePx, f32>>,
    pub running_animations: bool,
    pub children: Vec<FrameId>,
    pub is_private: bool,
}

A uniquely-identifiable pipeline of script thread, layout thread, and paint thread.

Fields
id: PipelineId parent_info: Option<(PipelineId, SubpageId, FrameType)> script_chan: IpcSender<ConstellationControlMsg> layout_chan: IpcSender<LayoutControlMsg>

A channel to layout, for performing reflows and shutdown.

compositor_proxy: Box<CompositorProxy + 'static + Send>

A channel to the compositor.

chrome_to_paint_chan: Sender<ChromeToPaintMsg> url: Url

URL corresponding to the most recently-loaded page.

title: Option<String>

The title of the most recently-loaded page.

size: Option<TypedSize2D<PagePx, f32>> running_animations: bool

Whether this pipeline is currently running animations. Pipelines that are running animations cause composites to be continually scheduled.

children: Vec<FrameId> is_private: bool
Methods

impl Pipeline

fn spawn<Message, LTF, STF>(state: InitialPipelineState) -> Result<(Pipeline, Option<ChildProcess>), IOError> where LTF: LayoutThreadFactory<Message=Message>, STF: ScriptThreadFactory<Message=Message>

Starts a paint thread, layout thread, and possibly a script thread, in a new process if requested.

fn new(id: PipelineId, parent_info: Option<(PipelineId, SubpageId, FrameType)>, script_chan: IpcSender<ConstellationControlMsg>, layout_chan: IpcSender<LayoutControlMsg>, compositor_proxy: Box<CompositorProxy + 'static + Send>, chrome_to_paint_chan: Sender<ChromeToPaintMsg>, url: Url, size: Option<TypedSize2D<PagePx, f32>>) -> Pipeline

fn grant_paint_permission(&self)

fn revoke_paint_permission(&self)

fn exit(&self)

fn freeze(&self)

fn thaw(&self)

fn force_exit(&self)

fn to_sendable(&self) -> CompositionPipeline

fn add_child(&mut self, frame_id: FrameId)

fn remove_child(&mut self, frame_id: FrameId)

fn trigger_mozbrowser_event(&self, subpage_id: SubpageId, event: MozBrowserEvent)